requirements.txt
aiohttp==3.11.2
async-timeout==5.0.1
charset-normalizer==3.4.0
discord.py==2.4.0
propcache==0.2.0
pycparser==2.22
requests==2.32.3
sympy==1.13.3
tabulate==0.9.0
typing_extensions==4.12.2
psutil==6.1.0
pypresence==4.3.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.26.1